requirements.txt
streamlit==1.31.0
openai==1.12.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2
tiktoken==0.7.0
httpx>=0.28.0
langdetect==1.0.9